src/requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
eventbrite-python==0.5.4
ticketmaster-python==0.1.0
montreal-api==0.1.0
pytz==2023.3
